pyproject.toml
[project]
name = "dbt-libsql"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yuta OHWATARI", email = "y.ohwatari@cartahd.com" }
]
dependencies = [
    "dbt-core~=1.5.0",
    "libsql-experimental>=0.0.34",
    "agate>=1.6, <1.7.1",
    "dbt-sqlite~=1.5.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "dbt_libsql:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.2",
    "pytest>=8.2.0",
    "mypy>=1.10.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_libsql"]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "ANN102", "EXE", "T20", "E501", "D"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
